python_services/requirements.txt
google-genai
requests
fastapi
uvicorn
pillow
torch
opencv-python
numpy
